Move celery to tasks
