+++ h/tasks/cleanup.py
@@ -4,7 +4,7 @@ import newrelic
 from sqlalchemy import func, select
 
 from h import models
-from h.celery import celery, get_task_logger
+from h.tasks.celery import celery, get_task_logger
 
 log = get_task_logger(__name__)
