+++ h/tasks/indexer.py
@@ -3,8 +3,8 @@ from abc import ABC
 import newrelic
 from celery import Task
 
-from h.celery import celery, get_task_logger
 from h.services import AnnotationSyncService
+from h.tasks.celery import celery, get_task_logger
 
 log = get_task_logger(__name__)
